requirements.txt
pytest==8.2.2
pytest-playwright==0.5.0
pytest-html==4.1.1
